fixed mask feature
git-svn-id: https://orchestra.med.harvard.edu/svn/lccb/matlab/projects/receptors/main@9906 0bcb941b-74d1-0310-bb26-e6636707f451

+++ ErbBproject/analysis/analyzeLMdataTracking.m
@@ -127,7 +127,7 @@ for iFile=1:nFiles
     if createMask
         % create ROI and store in MD
         img=double(MD.channels_(1).loadImage(1));
-        mask=roipoly(img/max(img(:)));
+        mask=double(roipoly(img/max(img(:))));
         close all;
         maskFile=MD.getFilename();
         maskFile(end-2:end)='tif';
